refactor: move club search helpers under util package
