+++ backend/src/main/java/moadong/club/util/search/EditDistanceCalculator.java
@@ -1,4 +1,4 @@
-package moadong.club.search;
+package moadong.club.util.search;
 
 import org.springframework.stereotype.Component;
