+++ backend/src/test/java/moadong/club/service/ClubSearchServiceTest.java
@@ -10,10 +10,10 @@ import java.util.Optional;
 import moadong.club.payload.dto.ClubSearchResult;
 import moadong.club.payload.response.ClubSearchResponse;
 import moadong.club.repository.ClubSearchRepository;
-import moadong.club.search.ClubSearchCandidate;
-import moadong.club.search.ClubSearchMatcher;
-import moadong.club.search.ClubSearchMatchType;
-import moadong.club.search.ClubSearchRanker;
+import moadong.club.util.search.ClubSearchCandidate;
+import moadong.club.util.search.ClubSearchMatcher;
+import moadong.club.util.search.ClubSearchMatchType;
+import moadong.club.util.search.ClubSearchRanker;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
