+++ backend/src/test/java/moadong/club/util/search/ClubSearchTextNormalizerTest.java
@@ -1,4 +1,4 @@
-package moadong.club.search;
+package moadong.club.util.search;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
